Creacion de las clases necesarias y correcction del controller
