+++ src/AppForSEII2526.API/Controllers/HerramientasController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 using AppForSEII2526.API.DTO.Alquilar_Herramienta;
+using AppForSEII2526.API.DTO.RepararDTOs;
 
 namespace AppForSEII2526.API.Controllers
 {
@@ -53,51 +54,7 @@ namespace AppForSEII2526.API.Controllers
 
 
         //Reparacion de herramientas Saelices
-        [HttpGet]
-        [Route("[action]")]
-        [ProducesResponseType(typeof(ReparacionDTO), (int)HttpStatusCode.OK)]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        public async Task<ActionResult> GetReparacion(int id)
-        {
-            if (_context.Reparacion == null)
-            {
-                _logger.LogError("Error: Reparaciones table does not exist");
-                return NotFound();
-            }
-
-            var reparacion = await _context.Reparacion
-                .Where(r => r.Id == id)
-                .Include(r => r.ItemsReparacion)         // join con ReparacionItem
-                    .ThenInclude(ri => ri.Herramienta)   // join con Herramienta
-                .Select(r => new ReparacionDTO(
-                    r.Id,
-                    r.Cliente,                           // Nombre completo del cliente
-                    r.FechaEntrega,
-                    r.FechaRecogida,
-                    r.PrecioTotal,
-                    r.ItemsReparacion
-                        .Select(ri => new ReparacionItemDTO(
-                            ri.Cantidad,
-                            ri.Descripcion,
-                            ri.Herramienta.Precio,
-                            ri.Herramienta.Id,
-                            ri.Herramienta,
-                            ri.reparacionId,
-                            ri.Reparacion
-
-                        )).ToList()
-                    , r.MetodoPago
-                ))
-                .FirstOrDefaultAsync();
-
-            if (reparacion == null)
-            {
-                _logger.LogWarning($"No se encontró la reparación con id {id}");
-                return NotFound();
-            }
-
-            return Ok(reparacion);
-        }
+        
 
         // Crear Oferta, Telmo
