+++ src/AppForSEII2526.API/DTO/RepararDTOs/ReparacionDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace AppForSEII2526.API.DTO
+﻿namespace AppForSEII2526.API.DTO.RepararDTOs
 {
     public class ReparacionDTO
     {
